code change in driver, division, terminal and iconstant file.

+++ DPU_UI/src/com/dpu/controller/TerminalEditController.java
@@ -26,7 +26,7 @@ public class TerminalEditController extends Application implements Initializable
 	@FXML
 	Button btnUpdateTerminal;
 
-	int terminalId = 0;
+	Long terminalId = 0l;
 
 	@FXML
 	TextField txtTerminalName, txtLocation;
